feat: move lib to utils
